+++ R/emuR-database.DBconfig.R
@@ -148,11 +148,12 @@ find_segmentLevels<-function(emuDBhandle, attrName){
   segLvlList=character(0)
   for(extLnkDef in extLnkDefs){
     if(extLnkDef[1]==lvlNm){
-      for(trgLvlNm in extLnkDef[2:length(extLnkDef)]){
-        
-        trgLd=get_levelDefinition(emuDBhandle, trgLvlNm)
-        if(trgLd['type']=='SEGMENT'){
-          segLvlList=unique(c(segLvlList,trgLvlNm))
+      if(length(extLnkDef) > 1){
+        for(trgLvlNm in extLnkDef[2:length(extLnkDef)]){
+          trgLd=get_levelDefinition(emuDBhandle, trgLvlNm)
+          if(trgLd['type']=='SEGMENT'){
+            segLvlList=unique(c(segLvlList,trgLvlNm))
+          }
         }
       }
     }
@@ -276,7 +277,7 @@ add_levelDefinition<-function(emuDBhandle, name,
   store_DBconfig(emuDBhandle, dbConfig)
   
   if(rewriteAllAnnots){
-    rewrite_allAnnots(emuDBhandle, verbose = verbose)
+    rewrite_annots(emuDBhandle, verbose = verbose)
   }
   invisible(NULL)
 }
@@ -379,7 +380,7 @@ remove_levelDefinition<-function(emuDBhandle, name, rewriteAllAnnots = TRUE, for
   store_DBconfig(emuDBhandle, dbConfig)
   
   if(rewriteAllAnnots){
-    rewrite_allAnnots(emuDBhandle, verbose = verbose)
+    rewrite_annots(emuDBhandle, verbose = verbose)
   }
   
   return(invisible(NULL))
@@ -509,7 +510,7 @@ internal_add_attributeDefinition <- function(emuDBhandle, levelName,
   store_DBconfig(emuDBhandle, dbConfig)
   
   if(rewriteAllAnnots){
-    rewrite_allAnnots(emuDBhandle, verbose = verbose)
+    rewrite_annots(emuDBhandle, verbose = verbose)
   }
   
 }
@@ -675,7 +676,7 @@ rename_attributeDefinition <- function(emuDBhandle, origAttrDef, newAttrDef, ver
   DBI::dbExecute(emuDBhandle$connection, paste0("DROP INDEX IF EXISTS level_rename_tmp_idx"))
   
   store_DBconfig(emuDBhandle, dbConfig)
-  rewrite_allAnnots(emuDBhandle, verbose = verbose)
+  rewrite_annots(emuDBhandle, verbose = verbose)
 }
 
 
@@ -749,7 +750,7 @@ remove_attributeDefinition <- function(emuDBhandle,
   # store changes
   store_DBconfig(emuDBhandle, dbConfig)
   if(rewriteAllAnnots){
-    rewrite_allAnnots(emuDBhandle, verbose = verbose)
+    rewrite_annots(emuDBhandle, verbose = verbose)
   }
 }
 
@@ -1234,7 +1235,7 @@ remove_linkDefinition <- function(emuDBhandle,
   # store changes
   store_DBconfig(emuDBhandle, dbConfig)
   if(force){
-    rewrite_allAnnots(emuDBhandle, verbose = verbose)
+    rewrite_annots(emuDBhandle, verbose = verbose)
   }
   
 }